Move proxystore.p2p.task to proxystore.utils.tasks
The proxystore.p2p.task module was not specific to the p2p module and
was used by modules outside of the p2p module. It makes more sense to
move this to a general utility module.
