+++ proxystore/p2p/relay/basic/client.py
@@ -34,8 +34,8 @@ from proxystore.p2p.relay.messages import RelayMessage
 from proxystore.p2p.relay.messages import RelayMessageDecodeError
 from proxystore.p2p.relay.messages import RelayRegistrationRequest
 from proxystore.p2p.relay.messages import RelayResponse
-from proxystore.p2p.task import spawn_guarded_background_task
 from proxystore.utils.environment import hostname
+from proxystore.utils.tasks import spawn_guarded_background_task
 
 logger = logging.getLogger(__name__)
