+++ proxystore/p2p/relay/basic/server.py
@@ -39,7 +39,7 @@ from proxystore.p2p.relay.messages import RelayMessageDecodeError
 from proxystore.p2p.relay.messages import RelayMessageEncodeError
 from proxystore.p2p.relay.messages import RelayRegistrationRequest
 from proxystore.p2p.relay.messages import RelayResponse
-from proxystore.p2p.task import spawn_guarded_background_task
+from proxystore.utils.tasks import spawn_guarded_background_task
 
 logger = logging.getLogger(__name__)
